ElixirCLRut (ElixirCLRUT v1.0.0)

A module that helps with validation and formatting of chilean RUT/RUN identifiers. Made by the Elixir Chile Community.

Functions

from(input)
(since 1.0.0)

@spec from(String.t()) :: struct()

@spec from(integer()) :: struct()

Standarizes the input and returns a struct that contains the formatted values.

This structure must be passed down to the validator function for checking if the RUT is valid.

Examples

iex> from("2228250-6")
%ElixirCLRut{
  from: "2228250-6",
  checkdigit: "6",
  formatted: "2.228.250-6",
  lastchar: "6",
  normalized: [2, 2, 2, 8, 2, 5, 0, 6]
}

iex> from("14193432")
%ElixirCLRut{
  from: "14193432",
  checkdigit: "5",
  formatted: "14.193.432-5",
  lastchar: "5",
  normalized: [1, 4, 1, 9, 3, 4, 3, 2]
}

valid?(input)
(since 1.0.0)

@spec valid?(struct()) :: {:ok, struct()} | {:error, struct()}

@spec valid?(String.t()) :: {:ok, struct()} | {:error, struct()}

@spec valid?(integer()) :: {:ok, struct()} | {:error, struct()}

Tells if the ElixirCLRut structure is valid. Beware that this will consider valid 1-9 and 0-0. Values normally considered fake ones and although valid, should be considered as not real values.

Returns a tuple:

{:ok, %ElixirCLRut{}} if valid

{:error, %ElixirCLRut{}} if is not valid.

Examples

iex> valid?("2228250-6")
{:ok, %ElixirCLRut{
  from: "2228250-6",
  checkdigit: "6",
  formatted: "2.228.250-6",
  lastchar: "6",
  normalized: [2, 2, 2, 8, 2, 5, 0, 6]
  }
}

iex> valid?("14193432")
{:ok, %ElixirCLRut{
  from: "14193432",
  checkdigit: "5",
  formatted: "14.193.432-5",
  lastchar: "5",
  normalized: [1, 4, 1, 9, 3, 4, 3, 2]
  }
}

iex> valid?(14193432)
{:ok, %ElixirCLRut{
  from: "14193432",
  checkdigit: "5",
  formatted: "14.193.432-5",
  lastchar: "5",
  normalized: [1, 4, 1, 9, 3, 4, 3, 2]
  }
}

iex> valid?(ElixirCLRut.from("14193432"))
{:ok, %ElixirCLRut{
  from: "14193432",
  checkdigit: "5",
  formatted: "14.193.432-5",
  lastchar: "5",
  normalized: [1, 4, 1, 9, 3, 4, 3, 2]
  }
}

valid?(input, bool, rules \\ [])
(since 1.0.0)

@spec valid?(struct(), boolean(), list()) :: {:ok, struct()} | {:error, struct()}

@spec valid?(struct(), boolean(), list()) :: {:ok, struct()} | {:error, struct()}

Strict mode will also validate format and avoid known fake ruts likes 1-9 and less than 6 length ruts. Some regex rules can be passed, if any of them match, it will consider the rut as invalid. strict false is the same as valid?/1.