ElixirCLRut.CheckDigit (ElixirCLRUT v1.0.0)

The Check Digit allows to verify if a RUT is correctly formed.

Functions

get(normalized)
(since 1.0.0)

@spec get(list()) :: String.t()

Calculates the check digit following the modulo 11 algorithm. Receives a list of numbers.

Examples

iex> get([2,2,2,8,2,5,0])
"6"

remove(normalized)
(since 1.0.0)

@spec remove(list()) :: list()

Removes the last item from the list. Used for cleaning up the check digit before the algorithm.

Examples

iex> remove([2,2,2,8,2,5,0,6])
[2,2,2,8,2,5,0]