ElixirCLRut.Struct (ElixirCLRUT v1.0.1)

A struct that holds information of the rut.

Functions

from(input, includes_checkdigit? \\ false)
(since 1.0.1)

@spec from(String.t(), boolean()) :: struct()

Standarizes the input and returns a struct that contains the formatted values.

This structure must be passed down to the validator function for checking if the RUT is valid.