ElixirCLRut.Validations (ElixirCLRUT v1.0.0)

Stores different validation functions

Functions

has_valid_checkdigit(token)
(since 1.0.0)

@spec has_valid_checkdigit(struct()) :: struct()

Receives a Token struct and validates checkdigit with lastdigit. If they are different it will add :wrong_checkdigit to the errors list.

length(token, length \\ 2)
(since 1.0.0)

@spec length(
  struct(),
  integer()
) :: struct()

Optional validation. Receives a Token struct and validates the normalized length is above or equal to the given length. Defaults to length 2.

Example

iex> (ElixirCLRut.validate("00-0") |> ElixirCLRut.Validations.length()).valid?
true

iex> (ElixirCLRut.validate("00-0") |> ElixirCLRut.Validations.length(3)).valid?
false

not_all_zeroes(token)
(since 1.0.0)

@spec not_all_zeroes(struct()) :: struct()

Optional validation. Receives a Token struct and validates that all chars are not jut zeroes.

Example

iex> (ElixirCLRut.validate("00000000-0") |> not_all_zeroes()).valid?
false

iex> (ElixirCLRut.validate("00000001-9") |> not_all_zeroes()).valid?
true

not_empty(token)
(since 1.0.0)

@spec not_empty(struct()) :: struct()

Receives a Token struct and counts the normalized array, if it's empty it will add :value_is_empty to the errors list.