ElixirCLRut (ElixirCLRUT v1.0.0)

A module that helps with validation and formatting of chilean RUT/RUN identifiers. Made by the Elixir Chile Community.

Functions

format(input, sep \\ ".")
(since 1.0.0)

@spec format(struct() | String.t(), String.t()) :: String.t()

Formats a string or Rut struct with the Rut format.

Examples

iex> format("1")
"1-9"

iex> format(from("6300948-2"), ",")
"6,300,948-2"

from(input)
(since 1.0.0)

@spec from(String.t()) :: struct()

When given a string it will return an ElixirCLRut.Struct

Example

iex> from("1-9")
%ElixirCLRut.Struct{checkdigit: "9", dashed?: true, from: "1-9", lastdigit: "9", normalized: [1], normalized_with_checkdigit: [1, 9]}

valid?(input)
(since 1.0.0)

@spec valid?(struct() | String.t()) :: boolean()

Performs a simple validation and returns the boolean result. true if is valid.

Examples

iex> valid?("1")
true

iex> valid?("6300948-1")
false

iex> valid?(from("6300948-1"))
false

validate(input)
(since 1.0.0)

@spec validate(struct() | String.t()) :: struct()

Executes the rut validations.

Examples

iex> validate("1").valid?
true

iex> validate("6300948-1").valid?
false